requirements.txt
opencv-python
easyocr
git+https://github.com/JaidedAI/EasyOCR.git

pymongo 
uvicorn
pymongo[srv] 
fastapi
python-multipart
Pillow
line-bot-sdk

httpx
passlib

PyJWT

bcrypt

pytz
